requirements.txt
streamlit
qdrant-client
pymupdf
python-docx
python-dotenv
requests
